+++ src/main/java/com/braintreegateway/ResourceCollection.java
@@ -8,13 +8,13 @@ import java.util.List;
 
 /**
  * A collection used to page through query or search results.
- * 
+ *
  * @param <T>
  *            type of object being paged, e.g. {@link Transaction} or
  *            {@link Customer}.
  */
 public class ResourceCollection<T> implements Iterable<T> {
-    
+
     private class PagedIterator<E> implements Iterator<E> {
         private ResourceCollection<E> resourceCollection;
         private List<String> ids;
@@ -22,7 +22,7 @@ public class ResourceCollection<T> implements Iterable<T> {
         private int index;
         private int nextIndexToFetch;
         private List<E> items;
-        
+
         public PagedIterator(ResourceCollection<E> resourceCollection) {
             this.resourceCollection = resourceCollection;
             this.ids = resourceCollection.ids;
@@ -31,7 +31,7 @@ public class ResourceCollection<T> implements Iterable<T> {
             this.nextIndexToFetch = 0;
             this.items = new ArrayList<E>();
         }
-        
+
         private List<String> nextBatchOfIds() {
             int lastIdIndex = nextIndexToFetch + pageSize;
             if (lastIdIndex > ids.size()) {
@@ -40,16 +40,16 @@ public class ResourceCollection<T> implements Iterable<T> {
 
             List<String> nextIds = ids.subList(nextIndexToFetch, lastIdIndex);
             nextIndexToFetch = lastIdIndex;
-            
+
             return nextIds;
         }
-        
+
         public boolean hasNext() {
             if (nextIndexToFetch < ids.size() && index == items.size()) {
                 this.items = resourceCollection.pager.getPage(nextBatchOfIds());
                 this.index = 0;
             }
-            
+
             if (index < items.size()) {
                 return true;
             }
@@ -76,10 +76,10 @@ public class ResourceCollection<T> implements Iterable<T> {
         pageSize = response.findInteger("page-size");
         ids = response.findAllStrings("ids/*");
     }
-    
+
     /**
      * Returns the approximate total size of the collection.
-     * 
+     *
      * @return Approximate size of collection
      */
     public int getMaximumSize() {
@@ -92,5 +92,9 @@ public class ResourceCollection<T> implements Iterable<T> {
 
     public T getFirst() {
         return pager.getPage(ids.subList(0, 1)).get(0);
+    }
+
+    public List<String> getIds() {
+        return ids;
     }
 }