+++ src/test/java/com/braintreegateway/integrationtest/ResourceCollectionIT.java
@@ -7,6 +7,7 @@ import com.braintreegateway.util.NodeWrapperFactory;
 import org.junit.Test;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import static org.junit.Assert.assertEquals;
@@ -49,5 +50,38 @@ public class ResourceCollectionIT {
             count++;
         }
         assertEquals(values.length, count);
+    }
+
+    @Test
+    public void getIds() {
+        NodeWrapper xml = NodeWrapperFactory.instance.create("<search-results>" +
+                "<page-size>2</page-size>" +
+                "<ids type=\"array\">" +
+                    "<items>0</items>" +
+                    "<items>1</items>" +
+                    "<items>2</items>" +
+                    "<items>3</items>" +
+                    "<items>4</items>" +
+                "</ids>" +
+                "</search-results>"
+        );
+
+        ResourceCollection<String> resourceCollection = new ResourceCollection<String>(new TestPager(), xml);
+
+        List<String> ids = Arrays.asList("0", "1", "2", "3", "4");
+        assertEquals(resourceCollection.getIds(), ids);
+    }
+
+    @Test
+    public void getIdsWithEmptyResourceCollection() {
+        NodeWrapper xml = NodeWrapperFactory.instance.create("<search-results>" +
+                "<page-size>2</page-size>" +
+                "</search-results>"
+        );
+
+        ResourceCollection<String> resourceCollection = new ResourceCollection<String>(new TestPager(), xml);
+
+        List<String> emptyIds = new ArrayList<String>();
+        assertEquals(resourceCollection.getIds(), emptyIds);
     }
 }